fix: thread-safe no_grad, math.prod, remove double astype
- tensor.py: replace global _grad_enabled with threading.local() so
  no_grad() context managers in different threads don't clobber each other
- flex/builder.py: ShapeTracer.flattened uses math.prod() instead of
  manual accumulation loop
- models/initialization.py: remove redundant double .astype(np.float32)
  in he_init()

+++ src/minicnn/models/initialization.py
@@ -72,7 +72,7 @@ def get_initializer(name: str):
 
 def he_init(size: int, fan_in: int, rng=None) -> np.ndarray:
     rng = rng or np.random.default_rng()
-    return (rng.standard_normal(size).astype(np.float32) * np.sqrt(2.0 / fan_in)).astype(np.float32)
+    return (rng.standard_normal(size) * np.sqrt(2.0 / fan_in)).astype(np.float32)
 
 
 def init_weights(seed: int, geom=None):
